Factor selector components out into their own files
